refactor: Request 객체 책임 분리
- Request는 HTTP 요청 값의 저장, 출력, 조회 역할만 담당하도록 변경
- RequestHandler는 HTTP 요청 파싱 및 Request 객체 생성 책임을 가지도록 수정
