+++ src/main/java/webserver/RequestHandler.java
@@ -1,34 +1,54 @@
 package webserver;
 
-import java.io.*;
-import java.net.Socket;
-
 import http.Request;
-import http.Response;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class RequestHandler implements Runnable {
-    private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
-    private Socket connection;
+public class RequestHandler {
+    private InputStream in;
 
-    public RequestHandler(Socket connectionSocket) {
-        this.connection = connectionSocket;
+    public RequestHandler(InputStream in) {
+        this.in = in;
     }
 
-    public void run() {
-        logger.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
-                connection.getPort());
+    public Request handleRequest() throws IOException {
+        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(in));
+        Map<String, String> requestLine = parseRequestLine(bufferedReader);
+        Map<String, String> headers = parseRequestHeaders(bufferedReader);
+        return new Request(requestLine, headers);
+    }
 
-        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
-            Request request = new Request(in);
-            request.printRequest();
-            String path = request.getRequestLine("path");
-            Response response = new Response(out, path);
-            response.sendResponse();
-        } catch (IOException e) {
-            logger.error(e.getMessage());
+    private Map<String, String> parseRequestLine(BufferedReader bufferedReader) throws IOException {
+        Map<String, String> requestLine = new LinkedHashMap<>();
+        String[] requestLineParts = bufferedReader.readLine().split(" ");
+        requestLine.put("method", requestLineParts[0]);
+        requestLine.put("path", requestLineParts[1]);
+        requestLine.put("protocol", requestLineParts[2]);
+        return requestLine;
+    }
+
+    private Map<String, String> parseRequestHeaders(BufferedReader bufferedReader) throws IOException {
+        Map<String, String> headers = new LinkedHashMap<>();
+
+        String line;
+
+        while ((line = bufferedReader.readLine()) != null) {
+
+            if (line.isEmpty()) break;
+
+            String[] headerParts = parseLine(line);
+            headers.put(headerParts[0], headerParts[1]);
         }
+
+        return headers;
+    }
+
+    private String[] parseLine(String line) {
+        return line.split(":");
     }
 }
